+++ docs/afoCoverage_testToAfos.txt
@@ -7,7 +7,6 @@ de.gematik.pki.pkits.testsuite.approval.CertificateApprovalTests
 de.gematik.pki.pkits.testsuite.approval.OcspApprovalTests
   verifyForCaCertificateWithoutServiceSupplyPoint    GS-A_4656, GS-A_4657
   verifyHashAlgorithmsInOcspSha1                     GS-A_4657, RFC 6960
-  verifyHashAlgorithmsInOcspSha2                     GS-A_4657, GS-A_4674, RFC 5754, RFC 6960
   verifyHashAlgorithmsInOcspShaVersion               GS-A_4657, GS-A_4674, RFC 6960
   verifyInvalidCerIdInOcspResponse                   GS-A_4657
   verifyInvalidCertHashInOcspResponse                GS-A_4657
@@ -18,18 +17,20 @@ de.gematik.pki.pkits.testsuite.approval.OcspApprovalTests
   verifyOcspGracePeriod                              A_22290, GS-A_4657, GS-A_4943
   verifyOcspRequestStructure                         GS-A_4657, GS-A_4674, RFC 5019, RFC 6960
   verifyOcspResponseMissingNextUpdate                GS-A_4657, GS-A_5215
+  verifyOcspResponseResponderIdByName                RFC 6960
+  verifyOcspResponseTimeoutAndDelay                  GS-A_4657
+  verifyOcspResponseVariousStatusAndResponseBytes    GS-A_4657
+  verifyOcspResponseWithNullParameterInCertId        RFC 5280, RFC 6960
+
+de.gematik.pki.pkits.testsuite.approval.OcspToleranceApprovalTests
   verifyOcspResponseNextUpdatePastOutOfTolerance     GS-A_4657, GS-A_5215
   verifyOcspResponseNextUpdatePastWithinTolerance    GS-A_4657, GS-A_5215
   verifyOcspResponseProducedAtFutureOutOfTolerance   GS-A_4657, GS-A_5215
   verifyOcspResponseProducedAtFutureWithinTolerance  GS-A_4657, GS-A_5215
   verifyOcspResponseProducedAtPastOutOfTolerance     GS-A_4657, GS-A_5215
   verifyOcspResponseProducedAtPastWithinTolerance    GS-A_4657, GS-A_5215
-  verifyOcspResponseResponderIdByName                RFC 6960
   verifyOcspResponseThisUpdateFutureOutOfTolerance   GS-A_4657, GS-A_5215
   verifyOcspResponseThisUpdateFutureWithinTolerance  GS-A_4657, GS-A_5215
-  verifyOcspResponseTimeoutAndDelay                  GS-A_4657
-  verifyOcspResponseVariousStatusAndResponseBytes    GS-A_4657
-  verifyOcspResponseWithNullParameterInCertId        RFC 5280, RFC 6960
 
 de.gematik.pki.pkits.testsuite.approval.TslApprovalExtraTests
   verifyExpiredTslInSystem  GS-A_4642, GS-A_4648, GS-A_4898, GS-A_5336
@@ -51,30 +52,32 @@ de.gematik.pki.pkits.testsuite.approval.TslApprovalTests
   verifyUseBackupTslDownload                             GS-A_4642, GS-A_4647, GS-A_4648
 
 de.gematik.pki.pkits.testsuite.approval.TslSignerApprovalTests
-  verifyForInvalidSignatureOfTslSigner                            GS-A_4642, GS-A_4648, GS-A_4650, GS-A_4655
-  verifyMissingOcspSignerInTslForTslSignerCert                    GS-A_4642, GS-A_4657
-  verifyOcspResponseTslSignerCertInvalidCertHash                  GS-A_4642, GS-A_4657
-  verifyOcspResponseTslSignerCertInvalidCertId                    GS-A_4642, GS-A_4657
-  verifyOcspResponseTslSignerCertMissingCertHash                  GS-A_4642, GS-A_4657
-  verifyOcspResponseTslSignerCertMissingNextUpdate                GS-A_4642, GS-A_4657, GS-A_5215
+  verifyForInvalidSignatureOfTslSigner                          GS-A_4642, GS-A_4648, GS-A_4650, GS-A_4655
+  verifyMissingOcspSignerInTslForTslSignerCert                  GS-A_4642, GS-A_4657
+  verifyOcspResponseTslSignerCertInvalidCertHash                GS-A_4642, GS-A_4657
+  verifyOcspResponseTslSignerCertInvalidCertId                  GS-A_4642, GS-A_4657
+  verifyOcspResponseTslSignerCertMissingCertHash                GS-A_4642, GS-A_4657
+  verifyOcspResponseTslSignerCertMissingNextUpdate              GS-A_4642, GS-A_4657, GS-A_5215
+  verifyOcspResponseTslSignerCertResponderIdByName              GS-A_4642, RFC 6960
+  verifyOcspResponseTslSignerCertStatusRevokedAndUnknown        GS-A_4642, GS-A_4657
+  verifyOcspResponseTslSignerCertTimeoutAndDelay                GS-A_4642, GS-A_4657
+  verifyOcspResponseTslSignerCertVariousStatusAndResponseBytes  GS-A_4642, GS-A_4657
+  verifyOcspResponseTslSignerCertWithNullParameterInCertId      GS-A_4642, RFC 5280
+  verifyOcspResponseWithInvalidSignatureForTslSignerCert        GS-A_4642, GS-A_4657
+  verifyTslSignerCertBroken                                     GS-A_4642
+  verifyTslSignerCertExpired                                    GS-A_4642, GS-A_4648, GS-A_4650, GS-A_4653
+  verifyTslSignerCertInvalidKeyUsageAndExtendedKeyUsage         GS-A_4642, GS-A_4648, GS-A_4650
+  verifyTslSignerCertNotYetValid                                GS-A_4642, GS-A_4648, GS-A_4650, GS-A_4653
+
+de.gematik.pki.pkits.testsuite.approval.TslSignerToleranceApprovalTests
   verifyOcspResponseTslSignerCertNextUpdatePastOutOfTolerance     GS-A_4642, GS-A_4657, GS-A_5215
   verifyOcspResponseTslSignerCertNextUpdatePastWithinTolerance    GS-A_4642, GS-A_4657, GS-A_5215
   verifyOcspResponseTslSignerCertProducedAtFutureOutOfTolerance   GS-A_4642, GS-A_4657, GS-A_5215
   verifyOcspResponseTslSignerCertProducedAtFutureWithinTolerance  GS-A_4642, GS-A_4657, GS-A_5215
   verifyOcspResponseTslSignerCertProducedAtPastOutOfTolerance     GS-A_4642, GS-A_4657, GS-A_5215
   verifyOcspResponseTslSignerCertProducedAtPastWithinTolerance    GS-A_4642, GS-A_4657, GS-A_5215
-  verifyOcspResponseTslSignerCertResponderIdByName                GS-A_4642, RFC 6960
-  verifyOcspResponseTslSignerCertStatusRevokedAndUnknown          GS-A_4642, GS-A_4657
   verifyOcspResponseTslSignerCertThisUpdateFutureOutOfTolerance   GS-A_4642, GS-A_4657, GS-A_5215
   verifyOcspResponseTslSignerCertThisUpdateFutureWithinTolerance  GS-A_4642, GS-A_4657, GS-A_5215
-  verifyOcspResponseTslSignerCertTimeoutAndDelay                  GS-A_4642, GS-A_4657
-  verifyOcspResponseTslSignerCertVariousStatusAndResponseBytes    GS-A_4642, GS-A_4657
-  verifyOcspResponseTslSignerCertWithNullParameterInCertId        GS-A_4642, RFC 5280
-  verifyOcspResponseWithInvalidSignatureForTslSignerCert          GS-A_4642, GS-A_4657
-  verifyTslSignerCertBroken                                       GS-A_4642
-  verifyTslSignerCertExpired                                      GS-A_4642, GS-A_4648, GS-A_4650, GS-A_4653
-  verifyTslSignerCertInvalidKeyUsageAndExtendedKeyUsage           GS-A_4642, GS-A_4648, GS-A_4650
-  verifyTslSignerCertNotYetValid                                  GS-A_4642, GS-A_4648, GS-A_4650, GS-A_4653
 
 de.gematik.pki.pkits.testsuite.approval.TslTaApprovalTests
   verifyHandlingOfStatusStartingTimeAndOverwriteAnnouncedInactiveTrustAnchors  GS-A_4642, GS-A_4643
