+++ pkits-common/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>de.gematik.pki.pkits</groupId>
     <artifactId>pki-testsuite</artifactId>
-    <version>3.0.2</version>
+    <version>4.0.0</version>
   </parent>
 
   <artifactId>pkits-common</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
   <name>Common Code</name>
   <description>Common code</description>
 
@@ -31,7 +31,7 @@
     </dependency>
     <dependency>
       <groupId>com.konghq</groupId>
-      <artifactId>unirest-java</artifactId>
+      <artifactId>unirest-java-core</artifactId>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
