+++ pkits-coverage-reports/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>de.gematik.pki.pkits</groupId>
     <artifactId>pki-testsuite</artifactId>
-    <version>3.0.2</version>
+    <version>4.0.0</version>
   </parent>
 
   <artifactId>pkits-coverage-reports</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
   <packaging>pom</packaging>
 
   <name>Code Coverage Reports</name>
