+++ pkits-ocsp-responder/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>de.gematik.pki.pkits</groupId>
     <artifactId>pki-testsuite</artifactId>
-    <version>3.0.2</version>
+    <version>4.0.0</version>
   </parent>
 
   <artifactId>pkits-ocsp-responder</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
   <name>OCSP Responder Simulator</name>
   <description>Spring Boot OCSP Responder</description>
 
@@ -69,7 +69,7 @@
 
     <dependency>
       <groupId>com.konghq</groupId>
-      <artifactId>unirest-java</artifactId>
+      <artifactId>unirest-java-core</artifactId>
     </dependency>
     <dependency>
       <groupId>jakarta.annotation</groupId>
