+++ pkits-testsuite/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>de.gematik.pki.pkits</groupId>
     <artifactId>pki-testsuite</artifactId>
-    <version>3.0.2</version>
+    <version>4.0.0</version>
   </parent>
 
   <artifactId>pkits-testsuite</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
   <name>PKI Test Suite</name>
   <description>PKI test suite</description>
 
@@ -26,9 +26,10 @@
     <testReportDirectory>${project.basedir}/../out/testreport</testReportDirectory>
 
     <!-- dependencies -->
-    <version.commons-lang3>3.14.0</version.commons-lang3>
-    <version.apache.sshd>2.13.1</version.apache.sshd>
-    <version.itextpdf.html2pdf>5.0.4</version.itextpdf.html2pdf>
+    <version.awaitility>4.2.0</version.awaitility>
+    <version.commons-lang3>3.16.0</version.commons-lang3>
+    <version.apache.sshd>2.13.2</version.apache.sshd>
+    <version.itextpdf.html2pdf>5.0.5</version.itextpdf.html2pdf>
     <version.jaxb-runtime>4.0.1</version.jaxb-runtime>
     <version.mockito>5.12.0</version.mockito>
     <version.picocli>4.7.6</version.picocli>
@@ -103,6 +104,7 @@
     <dependency>
       <groupId>org.awaitility</groupId>
       <artifactId>awaitility</artifactId>
+      <version>${version.awaitility}</version>
     </dependency>
 
     <dependency>
@@ -146,7 +148,7 @@
     <!-- required for mvn verify -->
     <dependency>
       <groupId>com.konghq</groupId>
-      <artifactId>unirest-java</artifactId>
+      <artifactId>unirest-java-core</artifactId>
     </dependency>
     <dependency>
       <groupId>com.ardikars.pcap</groupId>
