+++ pkits-testsuite/src/main/java/de/gematik/pki/pkits/testsuite/approval/TslSignerApprovalTests.java
@@ -37,7 +37,6 @@ import static de.gematik.pki.pkits.testsuite.usecases.UseCaseResult.USECASE_VALI
 import static java.lang.Integer.max;
 import static java.lang.Math.round;
 
-import de.gematik.pki.gemlibpki.ocsp.OcspConstants;
 import de.gematik.pki.gemlibpki.ocsp.OcspResponseGenerator.CertificateIdGeneration;
 import de.gematik.pki.gemlibpki.ocsp.OcspResponseGenerator.ResponderIdType;
 import de.gematik.pki.gemlibpki.utils.P12Container;
@@ -46,7 +45,6 @@ import de.gematik.pki.pkits.ocsp.responder.data.CustomCertificateStatusDto;
 import de.gematik.pki.pkits.ocsp.responder.data.CustomCertificateStatusType;
 import de.gematik.pki.pkits.ocsp.responder.data.OcspResponderConfig;
 import de.gematik.pki.pkits.ocsp.responder.data.OcspResponderConfig.OcspResponderConfigBuilder;
-import de.gematik.pki.pkits.testsuite.common.DtoDateConfigOption;
 import de.gematik.pki.pkits.testsuite.common.PkitsTestSuiteUtils;
 import de.gematik.pki.pkits.testsuite.common.tsl.TslDownload;
 import de.gematik.pki.pkits.testsuite.common.tsl.generation.TslDownloadGenerator;
@@ -158,236 +156,6 @@ class TslSignerApprovalTests extends ApprovalTestsBase {
 
     updateTrustStoreWithAlternativeCerts(
         dtoBuilder -> dtoBuilder.validSignature(false), TSL_UPDATE_NOT_EXPECTED, USECASE_INVALID);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_025 (Data Variant 4) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with producedAt in past within tolerance")
-  void verifyOcspResponseTslSignerCertProducedAtPastWithinTolerance() {
-
-    initialState();
-
-    final int producedAtDeltaMilliseconds =
-        -(OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS
-            - ocspSettings.getTimeoutDeltaMilliseconds());
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.PRODUCED_AT, producedAtDeltaMilliseconds),
-        TSL_UPDATE_EXPECTED,
-        USECASE_VALID);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_025 (Data Variant 3) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with producedAt in past out of tolerance")
-  void verifyOcspResponseTslSignerCertProducedAtPastOutOfTolerance() {
-
-    initialState();
-
-    final int producedAtDeltaMilliseconds =
-        -(OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS
-            + ocspSettings.getTimeoutDeltaMilliseconds());
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.PRODUCED_AT, producedAtDeltaMilliseconds),
-        TSL_UPDATE_NOT_EXPECTED,
-        USECASE_INVALID);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_025 (Data Variant 2) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with producedAt in future within tolerance")
-  void verifyOcspResponseTslSignerCertProducedAtFutureWithinTolerance() {
-
-    initialState();
-
-    final int producedAtDeltaMilliseconds =
-        OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS - ocspSettings.getTimeoutDeltaMilliseconds();
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.PRODUCED_AT, producedAtDeltaMilliseconds),
-        TSL_UPDATE_EXPECTED,
-        USECASE_VALID);
-
-    waitForOcspCacheToExpire(
-        testSuiteConfig.getTestObject().getOcspGracePeriodSeconds()
-            + producedAtDeltaMilliseconds / 1000);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_025 (Data Variant 1) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with producedAt in future out of tolerance")
-  void verifyOcspResponseTslSignerCertProducedAtFutureOutOfTolerance() {
-
-    initialState();
-
-    final int producedAtDeltaMilliseconds =
-        OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS
-            + ocspSettings.getTimeoutDeltaMilliseconds()
-            + 1000 * testSuiteConfig.getTestObject().getOcspGracePeriodSeconds();
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.PRODUCED_AT, producedAtDeltaMilliseconds),
-        TSL_UPDATE_NOT_EXPECTED,
-        USECASE_INVALID);
-
-    waitForOcspCacheToExpire(
-        testSuiteConfig.getTestObject().getOcspGracePeriodSeconds()
-            + producedAtDeltaMilliseconds / 1000);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_026 (Data Variant 2) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with thisUpdate in future within tolerance")
-  void verifyOcspResponseTslSignerCertThisUpdateFutureWithinTolerance() {
-
-    initialState();
-
-    final int thisUpdateDeltaMilliseconds =
-        OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS - ocspSettings.getTimeoutDeltaMilliseconds();
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.THIS_UPDATE, thisUpdateDeltaMilliseconds),
-        TSL_UPDATE_EXPECTED,
-        USECASE_VALID);
-
-    waitForOcspCacheToExpire(
-        testSuiteConfig.getTestObject().getOcspGracePeriodSeconds()
-            + thisUpdateDeltaMilliseconds / 1000);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_026 (Data Variant 1) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with thisUpdate in future out of tolerance")
-  void verifyOcspResponseTslSignerCertThisUpdateFutureOutOfTolerance() {
-
-    initialState();
-
-    final int thisUpdateDeltaMilliseconds =
-        OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS + ocspSettings.getTimeoutDeltaMilliseconds();
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.THIS_UPDATE, thisUpdateDeltaMilliseconds),
-        TSL_UPDATE_NOT_EXPECTED,
-        USECASE_INVALID);
-
-    waitForOcspCacheToExpire(
-        testSuiteConfig.getTestObject().getOcspGracePeriodSeconds()
-            + thisUpdateDeltaMilliseconds / 1000);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_029 (Data Variant 2) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with nextUpdate in past within tolerance")
-  void verifyOcspResponseTslSignerCertNextUpdatePastWithinTolerance() {
-
-    initialState();
-
-    final int nextUpdateAtDeltaMilliseconds =
-        -(OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS
-            - ocspSettings.getTimeoutDeltaMilliseconds());
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.NEXT_UPDATE, nextUpdateAtDeltaMilliseconds),
-        TSL_UPDATE_EXPECTED,
-        USECASE_VALID);
-
-    establishDefaultTrustStoreAndExecuteUseCase();
-  }
-
-  /** gematikId: UE_PKI_TC_0105_029 (Data Variant 1) */
-  @Test
-  @Afo(
-      afoId = "GS-A_4642",
-      description = "TUC_PKI_001: Periodische Aktualisierung TI-Vertrauensraum - Schritt 4")
-  @Afo(afoId = "GS-A_4657", description = "TUC_PKI_006: OCSP-Abfrage - Schritt 6")
-  @Afo(
-      afoId = "GS-A_5215",
-      description = "Festlegung der zeitlichen Toleranzen in einer OCSP-Response")
-  @DisplayName(
-      "Test OCSP response of TSL signer certificate with nextUpdate in past out of tolerance")
-  void verifyOcspResponseTslSignerCertNextUpdatePastOutOfTolerance() {
-
-    initialState();
-
-    final int nextUpdateDeltaMilliseconds =
-        -(OcspConstants.OCSP_TIME_TOLERANCE_MILLISECONDS
-            + ocspSettings.getTimeoutDeltaMilliseconds());
-
-    updateTrustStoreWithAlternativeCerts(
-        getDateConfigStep(DtoDateConfigOption.NEXT_UPDATE, nextUpdateDeltaMilliseconds),
-        TSL_UPDATE_NOT_EXPECTED,
-        USECASE_INVALID);
 
     establishDefaultTrustStoreAndExecuteUseCase();
   }
