+++ pkits-tls-client/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>de.gematik.pki.pkits</groupId>
     <artifactId>pki-testsuite</artifactId>
-    <version>3.0.2</version>
+    <version>4.0.0</version>
   </parent>
 
   <artifactId>pkits-tls-client</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
 
   <name>TLS Client</name>
   <description>TLS Client</description>
