+++ pkits-tls-client/src/main/java/de/gematik/pki/pkits/tls/client/TlsSettings.java
@@ -26,4 +26,5 @@ public class TlsSettings {
   String[] ecCiphersSuites;
   String[] rsaCiphersSuites;
   String[] enabledProtocols;
+  boolean sniEnabled;
 }
