+++ pkits-tls-client/src/main/resources/config.yml
@@ -9,3 +9,4 @@ tlsSettings:
   rsaCiphersSuites:
     - TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384
     - TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256
+  sniEnabled: true
