+++ pkits-tsl-generator/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>de.gematik.pki.pkits</groupId>
     <artifactId>pki-testsuite</artifactId>
-    <version>3.0.2</version>
+    <version>4.0.0</version>
   </parent>
 
   <artifactId>pkits-tsl-generator</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
   <name>PKI TSL Generator</name>
   <description>PKI TSL Generator</description>
