+++ pkits-tsl-provider/pom.xml
@@ -4,11 +4,11 @@
   <parent>
     <groupId>de.gematik.pki.pkits</groupId>
     <artifactId>pki-testsuite</artifactId>
-    <version>3.0.2</version>
+    <version>4.0.0</version>
   </parent>
 
   <artifactId>pkits-tsl-provider</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
   <name>TSL Server Simulator</name>
   <description>Spring Boot TSL Provider</description>
 
@@ -73,7 +73,7 @@
 
     <dependency>
       <groupId>com.konghq</groupId>
-      <artifactId>unirest-java</artifactId>
+      <artifactId>unirest-java-core</artifactId>
     </dependency>
     <dependency>
       <groupId>jakarta.annotation</groupId>
