+++ pkits-tsl-provider/src/test/java/de/gematik/pki/pkits/tsl/provider/api/TslProviderManagerTest.java
@@ -29,8 +29,8 @@ import de.gematik.pki.pkits.common.PkitsConstants;
 import de.gematik.pki.pkits.tsl.provider.data.TslRequestHistory;
 import de.gematik.pki.pkits.tsl.provider.data.TslRequestHistoryEntryDto;
 import java.util.List;
-import kong.unirest.HttpResponse;
-import kong.unirest.Unirest;
+import kong.unirest.core.HttpResponse;
+import kong.unirest.core.Unirest;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
