+++ pkits-tsl-provider/src/test/java/de/gematik/pki/pkits/tsl/provider/controller/TslInfoControllerTest.java
@@ -30,9 +30,9 @@ import de.gematik.pki.pkits.tsl.provider.common.TslConfigurator;
 import de.gematik.pki.pkits.tsl.provider.data.TslInfoRequestDto;
 import de.gematik.pki.pkits.tsl.provider.data.TslInfoRequestDto.HistoryDeleteOption;
 import java.nio.charset.StandardCharsets;
-import kong.unirest.HttpResponse;
-import kong.unirest.Unirest;
-import kong.unirest.UnirestException;
+import kong.unirest.core.HttpResponse;
+import kong.unirest.core.Unirest;
+import kong.unirest.core.UnirestException;
 import org.apache.http.HttpStatus;
 import org.assertj.core.api.Assertions;
 import org.json.JSONArray;
