+++ pkits-tsl-provider/src/test/java/de/gematik/pki/pkits/tsl/provider/controller/TslProviderControllerTest.java
@@ -29,8 +29,8 @@ import de.gematik.pki.pkits.tsl.provider.api.TslProviderManager;
 import de.gematik.pki.pkits.tsl.provider.data.TslProviderConfigDto;
 import de.gematik.pki.pkits.tsl.provider.data.TslProviderEndpointsConfig;
 import java.nio.charset.StandardCharsets;
-import kong.unirest.HttpResponse;
-import kong.unirest.Unirest;
+import kong.unirest.core.HttpResponse;
+import kong.unirest.core.Unirest;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
