+++ pom.xml
@@ -4,12 +4,12 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>3.3.1</version>
+    <version>3.3.2</version>
   </parent>
 
   <groupId>de.gematik.pki.pkits</groupId>
   <artifactId>pki-testsuite</artifactId>
-  <version>3.0.2</version>
+  <version>4.0.0</version>
   <packaging>pom</packaging>
   <name>PKI Testsuite</name>
   <description>PKI Testsuite to test products of the telematic infrastructure to fulfill the
@@ -85,28 +85,27 @@
 
     <!-- dependencies -->
     <version.ardikars.pcap>1.5.1</version.ardikars.pcap>
-    <version.assertj-core>3.26.0</version.assertj-core>
-    <version.awaitility>4.2.0</version.awaitility>
+    <version.assertj-core>3.26.3</version.assertj-core>
     <version.bouncycastle>1.78.1</version.bouncycastle>
-    <version.commons-compress>1.26.2</version.commons-compress>
+    <version.commons-compress>1.27.0</version.commons-compress>
     <version.commons-io>2.16.1</version.commons-io>
     <version.commons-text>1.12.0</version.commons-text>
     <version.jfiglet>1.0.1</version.jfiglet>
-    <version.gemLibPki>2.1.4</version.gemLibPki>
-    <version.jackson>2.17.1</version.jackson>
+    <version.gemLibPki>2.1.7</version.gemLibPki>
+    <version.jackson>2.17.2</version.jackson>
     <version.jakarta.annotation-api>3.0.0</version.jakarta.annotation-api>
-    <version.junit-jupiter>5.10.2</version.junit-jupiter>
-    <version.junit-platform-launcher>1.10.2</version.junit-platform-launcher>
+    <version.junit-jupiter>5.10.3</version.junit-jupiter>
+    <version.junit-platform-launcher>1.10.3</version.junit-platform-launcher>
     <version.log4j2>2.23.1</version.log4j2>
-    <version.lombok>1.18.32</version.lombok>
-    <version.openapi>2.5.0</version.openapi>
-    <version.unirest-java>3.14.5</version.unirest-java>
+    <version.lombok>1.18.34</version.lombok>
+    <version.openapi>2.6.0</version.openapi>
+    <version.unirest-java>4.4.4</version.unirest-java>
     <version.xmlunit>2.10.0</version.xmlunit>
 
     <!--  plugins -->
     <version.dependency-check-maven>7.2.1</version.dependency-check-maven>
-    <version.docker-maven-plugin>0.44.0</version.docker-maven-plugin>
-    <version.git-commit-id-maven-plugin>9.0.0</version.git-commit-id-maven-plugin>
+    <version.docker-maven-plugin>0.45.0</version.docker-maven-plugin>
+    <version.git-commit-id-maven-plugin>9.0.1</version.git-commit-id-maven-plugin>
     <version.jacoco-maven-plugin>0.8.12</version.jacoco-maven-plugin>
     <version.maven-assembly-plugin>3.6.0</version.maven-assembly-plugin>
     <version.maven-clean-plugin>3.4.0</version.maven-clean-plugin>
@@ -114,11 +113,11 @@
     <version.maven-deploy-plugin>3.1.2</version.maven-deploy-plugin>
     <version.maven-enforcer-plugin>3.5.0</version.maven-enforcer-plugin>
     <version.maven-failsafe-plugin>${version.maven-surefire-plugin}</version.maven-failsafe-plugin>
-    <version.maven-project-info-reports-plugin>3.6.1</version.maven-project-info-reports-plugin>
+    <version.maven-project-info-reports-plugin>3.6.2</version.maven-project-info-reports-plugin>
     <version.maven-resources-plugin>3.3.0</version.maven-resources-plugin>
-    <version.maven-site-plugin>3.12.1</version.maven-site-plugin>
+    <version.maven-site-plugin>3.20.0</version.maven-site-plugin>
     <version.maven-source-plugin>3.3.1</version.maven-source-plugin>
-    <version.maven-surefire-plugin>3.3.0</version.maven-surefire-plugin>
+    <version.maven-surefire-plugin>3.4.0</version.maven-surefire-plugin>
     <version.sonar-maven-plugin>4.0.0.4121</version.sonar-maven-plugin>
     <version.spotless-maven-plugin>2.43.0</version.spotless-maven-plugin>
     <version.spotless.google-java-format>1.17.0</version.spotless.google-java-format>
@@ -146,11 +145,6 @@
         <artifactId>assertj-core</artifactId>
         <version>${version.assertj-core}</version>
         <scope>test</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.awaitility</groupId>
-        <artifactId>awaitility</artifactId>
-        <version>${version.awaitility}</version>
       </dependency>
       <dependency>
         <groupId>org.bouncycastle</groupId>
@@ -311,7 +305,7 @@
       </dependency>
       <dependency>
         <groupId>com.konghq</groupId>
-        <artifactId>unirest-java</artifactId>
+        <artifactId>unirest-java-core</artifactId>
         <version>${version.unirest-java}</version>
       </dependency>
